requirements.txt
fastapi
uvicorn[standard]
scikit-learn
joblib
numpy
pydantic
python-dotenv
gunicorn
